全新版本，删除所有暂时不需要的东西
