+++ .settings/org.eclipse.core.resources.prefs
@@ -2,6 +2,5 @@ eclipse.preferences.version=1
 encoding//src/main/java=UTF-8
 encoding//src/main/resources=UTF-8
 encoding//src/main/resources/application.properties=UTF-8
-encoding//src/main/webapp/WEB-INF/jsp/test.jsp=UTF-8
 encoding//src/test/java=UTF-8
 encoding/<project>=UTF-8
