+++ src/main/java/course/controller/register/ParentUserRegisterController.java
@@ -1,4 +1,4 @@
-package course.controller;
+package course.controller.register;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -8,34 +8,35 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
 
-import course.bean.TeacherUser;
+import course.bean.ParentUser;
 import course.bean.User;
-import course.service.TeacherUserService;
+import course.service.ParentUserService;
+import util.RegisterTemplate;
 
 @Controller
-public class TeacherUserRegisterController extends RegisterTemplate
+class ParentUserRegisterController extends RegisterTemplate
 {
 	@Autowired
-	TeacherUserService tus;
+	ParentUserService pus;
 	
 	
 	/**
 	 * 处理注册请求
 	 */
-	@RequestMapping(value="/registerTeacherUser",method = RequestMethod.POST)
-	public ModelAndView registerPOST(@Validated TeacherUser user,BindingResult result)
+	@RequestMapping(value="/registerParentUser",method = RequestMethod.POST)
+	public ModelAndView registerPOST(@Validated ParentUser user,BindingResult result)
 	{
 		return template(user, result);
 	}
-	
+		
 	/**
 	 *  返回注册表单页面
 	 */
-	@RequestMapping(path = "/registerTeacherUser", method = RequestMethod.GET)
+	@RequestMapping(path = "/registerParentUser", method = RequestMethod.GET)
 	public ModelAndView registerGet(User user)
 	{
 		ModelAndView mav = new ModelAndView();
-		mav.setViewName("registerTeacherUser");
+		mav.setViewName("registerParentUser");
 		return mav;
 	}
 	
@@ -43,6 +44,7 @@ public class TeacherUserRegisterController extends RegisterTemplate
 	@Override
 	protected boolean registerUser(Object user)
 	{
-		return tus.register((TeacherUser)user);
+		return pus.register((ParentUser)user);
 	}
+	
 }
