+++ src/main/java/course/controller/user/ParentUserController.java
@@ -1,4 +1,4 @@
-package course.controller;
+package course.controller.user;
 
 import java.util.HashMap;
 import java.util.Map;
