+++ src/main/java/util/LoginTemplate.java
@@ -1,4 +1,4 @@
-package course.controller;
+package util;
 
 import java.util.HashMap;
 import java.util.Map;
