--- src/main/webapp/WEB-INF/jsp/test.jsp
@@ -1,24 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8" isELIgnored="false" %>
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="utf-8">
-<title>菜鸟教程(runoob.com)</title>
-<script src="https://cdn.staticfile.org/jquery/1.10.2/jquery.min.js">
-</script>
-<script src="/js/pop.min.js">
-</script>
-<script>
-function f(x)
-{
-	alert("#num"+iId.toString());
-}
-</script>
-</head>
-<body>
-
-<button onclick="f(x)"></button>
-
-</body>
-</html>
